% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{as.data.frame.rerand_design_result}
\alias{as.data.frame.rerand_design_result}
\title{Return design data with a treatment-assignment column}
\usage{
\method{as.data.frame}{rerand_design_result}(
  x,
  row.names = NULL,
  optional = FALSE,
  treatment = "Z",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{rerand_design_result} object.}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{treatment}{Name of the treatment-assignment column.}

\item{overwrite}{Whether to replace an existing column named \code{treatment}.}

\item{...}{Ignored.}
}
\description{
Return design data with a treatment-assignment column
}
