% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gen_X}
\alias{gen_X}
\title{Centered Covariate Generation Function}
\usage{
gen_X(n, K, Sigma_X = diag(K), seed = NULL)
}
\arguments{
\item{n}{Interger scalar; number of observations. n > 2.}

\item{K}{Integer scalar; number of covariates. K > 0.}

\item{Sigma_X}{Numeric matrix; K × K covariance matrix for X and must be positive semi definite. Default is identity matrix (independent covariates).}

\item{seed}{Integer scalar; random seed  reproducibility. Default is NULL (no seed set).}
}
\value{
Numeric matrix; n × K covariate matrix.
}
\description{
Generates n observations of K covariates drawn from a multivariate normal distribution with zero mean and specified covariance structure.
}
\examples{
set.seed(123)
X_data <- gen_X(n = 100, K = 5)

}
