% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{rerand_compare}
\alias{rerand_compare}
\title{Compare inference results from different estimators}
\usage{
rerand_compare(inferences)
}
\arguments{
\item{inferences}{A named collection created with \code{\link[=c.rerand_inference]{c.rerand_inference()}}.}
}
\value{
An object of class \code{rerand_compare}.
}
\description{
Compare inference results from different estimators
}
