% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{est_diff_unadj}
\alias{est_diff_unadj}
\title{Unadjusted Difference-in-Means Estimator Function}
\usage{
est_diff_unadj(Y_obs, Z, X = NULL)
}
\arguments{
\item{Y_obs}{Numeric vector; observed outcomes with length n.}

\item{Z}{Numeric vector; treatment assignment with length n (0 = control, 1 = treatment).}

\item{X}{Numeric matrix; n × K covariate matrix, used for constructing improved variance estimator. Optional. Default is NULL.}
}
\value{
A list containing:
\describe{
\item{tau_hat}{Numeric scalar; estimated ATE.}
\item{se_neyman}{Numeric scalar; standard error of the estimate using Neyman's conservative variance estimator.}
\item{se_ding}{Numeric scalar; standard error of the estimate using Ding's improved variance estimator.}
\item{method}{Character scalar; method used ("difference-in-means unadjusted").}
}
}
\description{
Calculates the unadjusted difference-in-means estimator for the average treatment effect (ATE) along with its standard error and 95\% confidence interval.
}
\examples{
res <- est_diff_unadj(Y_obs = rnorm(100), Z = rbinom(100, 1, 0.5))

}
