% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_utils.R
\name{get_quantile}
\alias{get_quantile}
\title{Generate Quantiles for Estimator under Rerandomization}
\usage{
get_quantile(R2, K, p_accept, alpha = 0.975, n_sim = 1e+05, engine = "cpp")
}
\arguments{
\item{R2}{Numeric scalar; the R-squared value representing the proportion of variance explained by covariates.}

\item{K}{Integer scalar; the number of covariates used in rerandomization.}

\item{p_accept}{Numeric scalar; the acceptance probability used in rerandomization.}

\item{alpha}{Numeric scalar; quantile level of mixed Gaussian distribution (default 0.975 for two-sided 95\% CI).}

\item{n_sim}{Integer scalar; number of simulation points to generate for estimating the quantiles (default 1e5).}

\item{engine}{Character scalar; computation engine to use, either "R" or "cpp". Default is "cpp".}
}
\value{
Numeric scalar; the estimated quantile of mixed Gaussian distribution.
}
\description{
This function simulates the distribution of the estimator under rerandomization to compute quantiles for confidence intervals. It accounts for the variance reduction due to rerandomization and the correlation structure of covariates.
}
\keyword{internal}
